Orpheus-FastAPI/requirements.txt
# Web Server Dependencies
fastapi==0.103.1
uvicorn==0.23.2
jinja2==3.1.2
pydantic==2.3.0
python-multipart==0.0.6

# API and Communication
requests==2.31.0
httpx==0.27.0
python-dotenv==1.0.0
watchfiles==1.0.4

# Audio Processing
numpy==1.24.0
sounddevice==0.4.6
snac==1.2.1       # Required for audio generation from tokens

# System Utilities
psutil==5.9.0

# PyTorch - Install separately with CUDA support:
# On Windows/Linux: 
#   pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
# On macOS:
#   pip3 install torch torchvision torchaudio

# Optional Dependencies
# For MP3 conversion (not currently implemented)
# pydub==0.25.1
# For better sentence splitting (potential future improvement)
# nltk==3.8.1

# Testing
pytest-asyncio==0.23.5
